+++ packages/seo/CHANGELOG.md
@@ -1,8 +1,18 @@
-# 3.0.0
+# Change Log
+
+## 4.0.0
+
+* Updated to use `angel3` packages
+* Published with `angel3` prefix
+
+## 3.0.0
+
 * Migrated to support Dart SDK 2.12.x NNBD
 
-# 2.0.0
+## 2.0.0
+
 * Angel 2 updates.
 
-# 1.0.0
+## 1.0.0
+
 * Initial release.